--- lib/src/home/card/card_bloc.dart
@@ -1,9 +1,0 @@
-import 'package:bloc_pattern/bloc_pattern.dart';
-
-class CardBloc extends BlocBase {
-  //dispose will be called automatically by closing its streams
-  @override
-  void dispose() {
-    super.dispose();
-  }
-}
