+++ lib/src/home/home_module.dart
@@ -1,6 +1,5 @@
 import 'package:perguntando/src/home/appbar/appbar_bloc.dart';
 import 'package:perguntando/src/home/bottom/bottom_bloc.dart';
-import 'package:perguntando/src/home/card/card_bloc.dart';
 import 'package:perguntando/src/home/background/background_bloc.dart';
 import 'package:perguntando/src/home/home_bloc.dart';
 import 'package:bloc_pattern/bloc_pattern.dart';
@@ -12,7 +11,6 @@ class HomeModule extends ModuleWidget {
   List<Bloc> get blocs => [
         Bloc((i) => AppbarBloc()),
         Bloc((i) => BottomBloc()),
-        Bloc((i) => CardBloc()),
         Bloc((i) => BackgroundBloc()),
         Bloc((i) => HomeBloc()),
       ];
